# ms-gateway/docs/swagger.yaml
definitions:
  helper.Message:
    properties:
      message:
        type: string
    type: object
  helper.MessageDetails:
    properties:
      detail:
        type: string
      message:
        type: string
    type: object
...
  model.Booking:
    properties:
      checkin_date:
        type: string
      checkout_date:
        type: string
      createdAt:
        type: string
      id:
        type: string
      room_Id:
        type: integer
      status:
        type: string
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
  model.CreateBookingRequest:
    properties:
      checkin_date:
        type: string
      checkout_date:
        type: string
      room:
        $ref: '#/definitions/model.RoomRequest'
      room_id:
        type: integer
    type: object
...
  model.RoomRequest:
    properties:
      capacity:
        type: string
      hotel_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: arviansyahnur3@gmail.com
    name: Arviansyah Nur
    url: http://www.swagger.io/support
  description: An e-commerce API for booking hotel and with many features, made using
    microservices and gRPC.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: vihotel
  version: "1.0"
paths:
  /api/v1/booking:
    get:
      consumes:
      - application/json
      description: Get all booking
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Booking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Booking
    post:
      consumes:
      - application/json
      description: Create booking
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Booking
  /api/v1/booking/:booking_id:
    delete:
      consumes:
      - application/json
      description: Update booking
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Booking
    get:
      consumes:
      - application/json
      description: Get booking
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Booking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Booking
    put:
      consumes:
      - application/json
      description: Update booking
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Booking
...
swagger: "2.0"